Fix circular channel creation
A plugin might listen for the "conversation-created" signal, and attempt
to write to the conversaton. The signal would be triggered in
purple_conversation_new, at which point haze_im_channel_constructor
hasn't finished, and therefore new_im_channel hasn't inserted the
conversation into the channels hashtable.

The result is a new attempt to create the channel and a segmentation
fault occurs.

This is triggered by msn-pecan which does exactly that to notify the
user that the contact doesn't have him on the contact list (reverse
subscription).

Signed-off-by: Felipe Contreras <felipe.contreras@gmail.com>
Signed-off-by: Will Thompson <will.thompson@collabora.co.uk>