Add only a single copy of a given debug message to TpDebugSender.
When HAZE_DEBUG_PURPLE was included in the debug arguments, both
g_log and log_to_debug_sender were being called. The log handler
was catching the g_log message and adding it to the debug sender
while log_to_debug_sender was also adding it to the debug sender.
The result was two debug messages being posted for all libpurple
debug messages. This patch calls one or the other, but not both,
depending on the debug flags.