Move 'rm -r' function to utility file.
In the process, relicense this function under a simple permissive
license. I am the copyright owner for this function, which was
introduced in commit 10e4006, and has sat unchanged except for
mechanical replacement of g_debug() with DEBUG() since then. This allows
this trivial utility code to be re-used in other glorious projects!