connection: generate user name from user splits
If a protocol has PurpleAccountUserSplits and "userplitX" parameters then merge
them together to generate the real user name for the Purple account.

This completes the PurpleAccountUserSplits support. Now it should be possible
for Maemo/MeeGo Account Setup UI to show multiple entry fields instead of
hacking around the problem that "," is not a legal character.