Mangle test name into connection parameters.
In Gabble, we push the name of the test being run into the connection
parameters by abusing the fact that Gabble accepts a resource as part of
the 'account' parameter, but overrides it with the value of the
'resource' parameter. Haze doesn't have a resource parameter, so we
can't use that trick… so, instead, let's shove it into the file transfer
proxy list parameter.